backend/requirements.txt
Flask==2.3.2
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
cohere==4.50.0
python-dotenv==1.0.1
duckduckgo-search
bleach==6.1.0
torch==2.3.0
transformers==4.41.1
flask-cors
playwright
google-search-results
